Update PKESK v6 to the latest version of the crypto refresh
Implement the changes from !304 of the crypto refresh:
https://gitlab.com/openpgp-wg/rfc4880bis/-/merge_requests/304

"Introduces an explicit length octet for the fingerprint.
This length also covers the key version number.  This has the
advantage of nicely covering the anonymous recipient case, where a
zero length indicates not having a fingerprint in the packet without
having to resort to a fictitious key version zero."